% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.regsurv}
\alias{coef.regsurv}
\title{Return model coefficients for a regsurv model}
\usage{
\method{coef}{regsurv}(object, s = NULL, ...)
}
\arguments{
\item{object}{regsurv object}

\item{s}{lambda value}

\item{...}{other arguments}
}
\value{
coefficients for a fitted regsurv model at all lambda's (s=NULL) or a specific value of lambda (s)
}
\description{
Return model coefficients for a regsurv model
}
