% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.regsurv}
\alias{cv.regsurv}
\title{Title}
\usage{
cv.regsurv(
  object,
  prep,
  nfolds = 10,
  plot = FALSE,
  force.nnhazards = TRUE,
  print = FALSE,
  feastol = 1e-08,
  ...
)
}
\arguments{
\item{object}{an object of class regsurv}

\item{prep}{an object of class survprep}

\item{nfolds}{number of cross-validation folds}

\item{plot}{plots deviance versus log(lambda) if TRUE}

\item{force.nnhazards}{the force.nnhazards status as used for regsurv() is always maintained during cross-validation, but the
force.nnhazards parameter in cv.regsurv() allows to force these constraints on the whole sample (TRUE) (as opposed to just
the in-fold cases (when FALSE)).}

\item{print}{prints progress if TRUE}

\item{feastol}{the tolerance on the primal and dual residual, default 1e-8}

\item{...}{other parameters (only used by internal functions)}
}
\value{
\item{oosll}{out-of-sample log-likelihood}
\item{cvm}{mean oosll per lambda}
\item{cvsd}{sd of oosll per lambda}
\item{msdr}{mean squared deviance residuals per lambda}
\item{lambda.grid}{sed grid of lambda values (taken to be the same as those used to fit to regsurv model)}
\item{lambda.min}{lambda value minimizing the (out-of-sample) deviance}
\item{lambda.min.index}{index for the value of lambda minimizing the (out-of-sample) deviance}
}
\description{
Title
}
