% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.regsurv}
\alias{plot.cv.regsurv}
\title{Plot cv.regsurv object}
\usage{
\method{plot}{cv.regsurv}(x, ...)
}
\arguments{
\item{x}{cv.regsurv object}

\item{...}{.. additional arguments that are passed along to plot()}
}
\value{
Nothing. Side-effect: plot.
}
\description{
Plot cv.regsurv object
}
