% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regsurv}
\alias{plot.regsurv}
\title{Plot coefficients stored in a regsurv object}
\usage{
\method{plot}{regsurv}(x, incl.baseline = FALSE, scaled.betas = FALSE, ...)
}
\arguments{
\item{x}{regsurv object}

\item{incl.baseline}{whether to include the parameters that relate to the baseline hazard or baseline cumulative hazard}

\item{scaled.betas}{plots coefficients for scaled columns of the design matrix if TRUE}

\item{...}{additional arguments that are passed along to matplot()}
}
\value{
Nothing. Side-effect: plot.
}
\description{
Plot coefficients stored in a regsurv object
}
