% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.regsurv}
\alias{predict.regsurv}
\title{Predictions for a regsurv model}
\usage{
\method{predict}{regsurv}(
  object,
  prep,
  lambda.index,
  newdata = NULL,
  type = c("cumhazard", "surv"),
  ...
)
}
\arguments{
\item{object}{regsurv object}

\item{prep}{survprep object used to fit the regsurv model}

\item{lambda.index}{lambda index for the required predictions (e.g. optimal value based on cross-validation)}

\item{newdata}{if provided, this should be a matrix with untransformed time-to-event in the first column and the model.matrix for
for main main predictor effects as would be provided to survprep()}

\item{type}{type of prediction ("cumhazard" or "surv")}

\item{...}{as for predict()}
}
\value{
prediction of the requested type
}
\description{
Predictions for a regsurv model
}
