% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.R
\name{rcs}
\alias{rcs}
\title{Restricted cubic splines basis functions}
\usage{
rcs(x, knots = stats::quantile(x, probs = c(0, 0.1, 0.5, 0.9, 1)))
}
\arguments{
\item{x}{numeric}

\item{knots}{numeric}
}
\value{
matrix of basis columns
}
\description{
Takes a numeric vector and a sequence of knots to generate restricted cubic splines basis
columns based on a truncated power series (as described in section 2.4.5 of the second edition of
Regression Modeling Strategies by Frank E. Harrrell, Jr).
}
\examples{
rcs(1:25)
}
