% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data for illustrative purposes}
\format{
A data frame with 250 rows and 12 columns
\describe{
\item{id}{identifier}
\item{eventtime}{eventtime}
\item{status}{status indicator (1=event, 0=censored)}
\item{x1}{covariates 1}
\item{x2}{covariates 2}
\item{x3}{covariates 3}
\item{x4}{covariates 4}
\item{x5}{covariates 5}
\item{x6}{covariates 6}
\item{x7}{covariates 7}
\item{x8}{covariates 8}
\item{x9}{covariates 9}
}
}
\usage{
simdata
}
\description{
Data were simulated using the simsurv package. The data generating mechanism is Weibull with
administrative censoring at time=10. The first 4 covariates have non-zero effects and the
remainder are noise variables. The first 2 covariate effects are time-varying. Details on the
data generating mechanism can be found in the source files (data-raw folder).
}
\keyword{datasets}
