% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survprep.R
\name{survprep}
\alias{survprep}
\title{Prepare data for regsurv()}
\usage{
survprep(
  tte,
  delta,
  X,
  model.scale = c("loghazard", "logHazard"),
  time.scale = c("logtime", "time"),
  spline.type = c("rcs", "ns"),
  ntimebasis,
  time.knots = NULL,
  tv = NULL,
  nitimebasis = NULL,
  itime.knots = NULL,
  qpoints = 9
)
}
\arguments{
\item{tte}{numeric vector containing time-to-event data}

\item{delta}{numeric vector containing event indicator data (1=event, 0=censored)}

\item{X}{design matrix containing all necessary covariate data}

\item{model.scale}{choose between log hazard and log cumulative hazard}

\item{time.scale}{choose between log time and linear time for the time scale of the model}

\item{spline.type}{type of spline used to model time dependencies (rcs=restricted cubic splines,
ns=natural splines (as implemented in the splines2 package))}

\item{ntimebasis}{integer number of basis columns for the spline representation of the log baseline hazard
or log cumulative baseline hazard.}

\item{time.knots}{numeric vector containing baseline spline knot locations.
Defaults to equally spaced knots (if NULL).}

\item{tv}{numeric vector of column indices for the columns in X with time-varying (i.e. non-proportional) effects.}

\item{nitimebasis}{number of basis columns for the spline representation of non-proportional
covariate effects.}

\item{itime.knots}{numeric vector containing spline knot locations for time-varying covariate
effects. Defaults to equally spaced knots (if NULL).}

\item{qpoints}{number of quadrature points for numerical integration over time (only applicable
for models on the log hazard scale)}
}
\value{
an object of class survprep
}
\description{
Prepare data for regsurv()
}
